# swagger/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/topicos/{id}":{"get":{"tags":["controller-topico"],"operationId":"carregarTopico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]},"put":{"tags":["controller-topico"],"operationId":"editarTopico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosEditarTopico"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/topicos/{id}/fechar":{"put":{"tags":["controller-topico"],"operationId":"fecharTopico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/topicos/{id}/desativar":{"put":{"tags":["controller-topico"],"operationId":"desativarTopico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/topicos/{id}/abrir":{"put":{"tags":["controller-topico"],"operationId":"abrirTopico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/usuarios":{"post":{"tags":["controller-usuario"],"operationId":"criarUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCriarUsuario"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/topicos":{"get":{"tags":["controller-topico"],"operationId":"carregarTopicos","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageDadosListagemTopicos"}}}}},"security":[{"bearer-key":[]}]},"post":{"tags":["controller-topico"],"operationId":"criarTopico","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCriarTopico"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/respostas":{"post":{"tags":["controller-resposta"],"operationId":"criarResposta","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCriarResposta"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/perfis":{"post":{"tags":["controller-perfil"],"operationId":"criarPerfil","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCriarPerfil"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/login":{"post":{"tags":["controller-autenticacao"],"operationId":"efetuarLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosAutenticacao"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/cursos":{"post":{"tags":["controller-curso"],"operationId":"criarCurso","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCriarCurso"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}},"/topicos/{id}/deletar":{"delete":{"tags":["controller-topico"],"operationId":"deletarTopico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearer-key":[]}]}}},"components":{"schemas":{"DadosEditarTopico":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"mensagem":{"type":"string"}},"required":["id","mensagem","titulo"]},"DadosCriarUsuario":{"type":"object","properties":{"nome":{"type":"string"},"email":{"type":"string"},"senha":{"type":"string"}},"required":["email","nome","senha"]},"DadosCriarTopico":{"type":"object","properties":{"titulo":{"type":"string"},"mensagem":{"type":"string"},"dataCriacao":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["ABERTO","FECHADO","DESATIVADO"]},"idUsuario":{"type":"integer","format":"int64"},"categoria":{"type":"string","enum":["FRONT_END","BACK_END","FULLSTACK","SOFTWARE_ENGINEERING","MOBILE","DATA_SCIENCE","DATA_ENGINEERING"]}},"required":["categoria","dataCriacao","idUsuario","mensagem","status","titulo"]},"DadosCriarResposta":{"type":"object","properties":{"mensagem":{"type":"string"},"idTopico":{"type":"integer","format":"int64"},"dataCriacao":{"type":"string","format":"date-time"},"idUsuario":{"type":"integer","format":"int64"}},"required":["dataCriacao","idTopico","idUsuario","mensagem"]},"DadosCriarPerfil":{"type":"object","properties":{"nome":{"type":"string"}},"required":["nome"]},"DadosAutenticacao":{"type":"object","properties":{"email":{"type":"string"},"senha":{"type":"string"}}},"DadosCriarCurso":{"type":"object","properties":{"nome":{"type":"string"},"categoriaCurso":{"type":"string","enum":["FRONT_END","BACK_END","FULLSTACK","SOFTWARE_ENGINEERING","MOBILE","DATA_SCIENCE","DATA_ENGINEERING"]}},"required":["categoriaCurso","nome"]},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"Curso":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"categoriaCurso":{"type":"string","enum":["FRONT_END","BACK_END","FULLSTACK","SOFTWARE_ENGINEERING","MOBILE","DATA_SCIENCE","DATA_ENGINEERING"]}}},"DadosListagemTopicos":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"mensagem":{"type":"string"},"dataCriacao":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["ABERTO","FECHADO","DESATIVADO"]},"usuario":{"$ref":"#/components/schemas/Usuario"},"curso":{"$ref":"#/components/schemas/Curso"},"respostas":{"type":"array","items":{"$ref":"#/components/schemas/Resposta"}}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"PageDadosListagemTopicos":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/DadosListagemTopicos"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"Perfil":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"usuarios":{"type":"array","items":{"$ref":"#/components/schemas/Usuario"}}}},"Resposta":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"mensagem":{"type":"string"},"topico":{"$ref":"#/components/schemas/Topico"},"dataCriacao":{"type":"string","format":"date-time"},"usuario":{"$ref":"#/components/schemas/Usuario"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"Topico":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"mensagem":{"type":"string"},"dataCriacao":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["ABERTO","FECHADO","DESATIVADO"]},"usuario":{"$ref":"#/components/schemas/Usuario"},"curso":{"$ref":"#/components/schemas/Curso"},"respostas":{"type":"array","items":{"$ref":"#/components/schemas/Resposta"}}}},"Usuario":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"senha":{"type":"string"},"perfis":{"type":"array","items":{"$ref":"#/components/schemas/Perfil"}},"ativo":{"type":"boolean"},"enabled":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"username":{"type":"string"},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"password":{"type":"string"}}}},"securitySchemes":{"bearer-key":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}